LayoutTests/imported/w3c/web-platform-tests/dom/events/EventTarget-dispatchEvent-expected.txt

CONSOLE MESSAGE: line 20: TypeError: aliases.forEach is not a function. (In 'aliases.forEach', 'aliases.forEach' is undefined)

PASS Calling dispatchEvent(null). 
